Add files via upload
Our daily operations related to watering plants are a major task in our routine.
It’s a cultural basis practice that is carried forward and most labor-intensive work.
Whatever might be the season, whichever day will be there at any weather watering
plants is done on a priority basis. Most urban areas of them have mini gardens,
Taking this idea in mind evolved into an automated water dispenser system. Which
can reduce manual efforts and the work can go in a simulated way. This system is
also connected to an app which is connected to a hardware system resulting in a
distribution of water to the plants. Gardners of small scale and wells as large scale,
for watering fruits and vegetables used can be addressed.